-Changes to classes and interfaces:     -ProgrammingFile: changed LOC from int to long     -IMeasurerService:         +GetProjects now returns a Solution         +Measure now takes a Solution as Input     -MeasurerService:         +GetProjects is now implemented         +Measure is now implemented     -ProjectIdentificatorUtils         +EnumPlatform is now private         +ItsAValidFile is now private         +WellSmellsLikeFile is now private -Solved the filter problem with Open button

+++ Measurer4000/Models/ProgrammingFile.cs
@@ -4,7 +4,7 @@
     {
         public string Name { get; set; }
 
-        public int LOC { get; set; }
+        public long LOC { get; set; }
 
         public bool IsUserInterface { get; set; }
